Move `init-pokemon-evolutions.ts` to `src/data/init/`
