+++ src/data/challenge.ts
@@ -1,29 +1,27 @@
-import type { BooleanHolder, NumberHolder } from "#app/utils";
-import { randSeedItem } from "#app/utils";
-import i18next from "i18next";
-import type { DexAttrProps, GameData } from "#app/system/game-data";
+import type { FixedBattleConfig } from "#app/battle";
 import { defaultStarterSpecies } from "#app/data/default-starters";
+import { pokemonEvolutions } from "#app/data/init/init-pokemon-evolutions";
+import { pokemonFormChanges } from "#app/data/pokemon-forms";
 import type PokemonSpecies from "#app/data/pokemon-species";
-import { getPokemonSpeciesForm } from "#app/utils/pokemon-species-utils";
-import { getPokemonSpecies } from "../utils/pokemon-species-utils";
 import { speciesStarterCosts } from "#app/data/starters";
 import type { Pokemon } from "#app/field/pokemon";
 import { PokemonMove } from "#app/field/pokemon-move";
-import type { FixedBattleConfig } from "#app/battle";
-import { BattleType } from "#enums/battle-type";
 import Trainer from "#app/field/trainer";
-import { TrainerVariant } from "#enums/trainer-variant";
 import type { GameMode } from "#app/game-mode";
-import { ElementalType } from "#enums/elemental-type";
+import type { DexAttrProps, GameData } from "#app/system/game-data";
+import { randSeedItem, type BooleanHolder, type NumberHolder } from "#app/utils";
+import { getPokemonSpecies, getPokemonSpeciesForm } from "#app/utils/pokemon-species-utils";
+import { BattleType } from "#enums/battle-type";
 import { Challenges } from "#enums/challenges";
+import { TypeColor, TypeShadowColor } from "#enums/color";
+import { ElementalType } from "#enums/elemental-type";
+import type { MoveId } from "#enums/move-id";
+import type { MoveSourceType } from "#enums/move-source-type";
+import { Nature } from "#enums/nature";
 import { SpeciesId } from "#enums/species-id";
 import { TrainerType } from "#enums/trainer-type";
-import { Nature } from "#enums/nature";
-import type { MoveId } from "#enums/move-id";
-import { TypeColor, TypeShadowColor } from "#enums/color";
-import { pokemonEvolutions } from "#app/data/pokemon-evolutions/init-pokemon-evolutions";
-import { pokemonFormChanges } from "#app/data/pokemon-forms";
-import type { MoveSourceType } from "#enums/move-source-type";
+import { TrainerVariant } from "#enums/trainer-variant";
+import i18next from "i18next";
 
 /** A constant for the default max cost of the starting party before a run */
 const DEFAULT_PARTY_MAX_COST = 10;
