+++ src/data/trainer-configs/rival-trainer-configs.ts
@@ -1,4 +1,4 @@
-import { pokemonEvolutions } from "#app/data/pokemon-evolutions/init-pokemon-evolutions";
+import { pokemonEvolutions } from "#app/data/init/init-pokemon-evolutions";
 import { pokemonPreEvolutions } from "#app/data/pokemon-pre-evolutions";
 import type PokemonSpecies from "#app/data/pokemon-species";
 import {
@@ -8,11 +8,11 @@ import {
   trainerPartyTemplates,
   type TrainerConfigs,
 } from "#app/data/trainer-config";
-import { TrainerSlot } from "#enums/trainer-slot";
 import type { PersistentModifier } from "#app/modifier/modifier";
 import { modifierTypes } from "#app/modifier/modifier-types";
 import { PokeballType } from "#enums/pokeball-type";
 import { SpeciesId } from "#enums/species-id";
+import { TrainerSlot } from "#enums/trainer-slot";
 import { TrainerType } from "#enums/trainer-type";
 
 let t = TrainerType.RIVAL;
